feat: Add API endpoint for deleting a user
This commit adds a new API endpoint in the `deleteUser.ts` file to handle the deletion of a user. It receives the `employee_id` from the request body and deletes the corresponding user from the 'voters' table in the database. If the user is not found or an error occurs during the deletion, an appropriate response is sent. This addition enhances the functionality of the application.

Fixes #123