feat: Add API endpoint for fetching candidates and voters
This commit adds a new API endpoint in the `getData.js` file to handle the fetching of candidates and voters from the database. It uses the `superbase` client to retrieve the data from the respective tables. If there is an error during the fetch, an appropriate response is sent. This addition enhances the functionality of the application.