feat: Add AddUser component
This commit adds a new component called AddUser, which is responsible for rendering a form to add a new user. The component includes input fields for employee ID and employee name, and a submit button to add the user. Upon successful addition, a success toast notification is displayed. This addition enhances the functionality of the application by providing a user-friendly way to add new users.

Fixes #N/A