feat: Add API endpoint for adding employees
The commit adds a new API endpoint in the `addEmployees.ts` file to handle the addition of employees to the database. It receives the employee ID and name from the request body and inserts them into the `voters` table using the `superbase` client. If the employee already exists, an error message is returned. This addition enhances the functionality of the application by providing a way to add employees.

Fixes #N/A