feat: Add API endpoint for verifying admin password
This commit adds a new API endpoint in the `verify-admin.ts` file to handle the verification of the admin password. It receives the password from the request body and queries the `admin` table in the database using the `superbase` client. If the password matches an entry in the table, a success response is returned. Otherwise, an error response is returned. This addition enhances the functionality of the application by providing a secure way to verify the admin password.

Fixes #N/A